Extension controller + service - int id to long + saving files changed with the new logic

+++ server-side/src/main/java/com/tick42/quicksilver/services/base/ExtensionService.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface ExtensionService {
 
-    Extension findById(int id, UserDetails loggedUser);
+    Extension findById(long id, UserDetails loggedUser);
 
     Extension update(Extension extension);
